script contents removed.
Too many repos were referring back to this sript and wasting GitHub actions minutes since this script can take up to 24 minutes to run. Its behavieor is no longer needed, but the myriad calling workflows still needed it to exist and succeed so as to not impact their existing behavior or cause failures in their own workflows.

Leaving in place but removing contents.